create composable for dialog
